IOS-512 Changed double to NSUInteger

+++ src/Catty/DataModel/Formula/FormulaElement.m
@@ -469,17 +469,17 @@
         }
         case MULTI_FINGER_TOUCHED: {
             BOOL screenTouched = [[TouchHandler shared] screenIsTouched];
-            BOOL isRequestedTouch = ([[TouchHandler shared] numberOfTouches] == left);
+            BOOL isRequestedTouch = ([[TouchHandler shared] numberOfTouches] == (NSUInteger)left);
             
             result = screenTouched && isRequestedTouch ? [NSNumber numberWithInt:1] : [NSNumber numberWithInt:0];
             break;
         }
         case MULTI_FINGER_X: {
-            result = [NSNumber numberWithFloat:[[TouchHandler shared] getPositionInSceneForTouchNumber:left].x];
+            result = [NSNumber numberWithFloat:[[TouchHandler shared] getPositionInSceneForTouchNumber:(NSUInteger)left].x];
             break;
         }
         case MULTI_FINGER_Y: {
-            result = [NSNumber numberWithFloat:[[TouchHandler shared] getPositionInSceneForTouchNumber:left].y];
+            result = [NSNumber numberWithFloat:[[TouchHandler shared] getPositionInSceneForTouchNumber:(NSUInteger)left].y];
             break;
         }
         default:
